+++ php/src/IcePHP/Connection.cpp
@@ -61,7 +61,7 @@ ZEND_METHOD(Ice_Connection, __toString)
     try
     {
         string str = _this->toString();
-        RETURN_STRINGL(STRCAST(str.c_str()), str.length(), 1);
+        RETURN_STRINGL(STRCAST(str.c_str()), static_cast<int>(str.length()), 1);
     }
     catch(const IceUtil::Exception& ex)
     {
