+++ php/src/IcePHP/Properties.cpp
@@ -62,7 +62,7 @@ ZEND_METHOD(Ice_Properties, __toString)
             }
             str.append(p->first + "=" + p->second);
         }
-        RETURN_STRINGL(STRCAST(str.c_str()), str.length(), 1);
+        RETURN_STRINGL(STRCAST(str.c_str()), static_cast<int>(str.length()), 1);
     }
     catch(const IceUtil::Exception& ex)
     {
@@ -88,7 +88,7 @@ ZEND_METHOD(Ice_Properties, getProperty)
     try
     {
         string val = _this->getProperty(propName);
-        RETURN_STRINGL(STRCAST(val.c_str()), val.length(), 1);
+        RETURN_STRINGL(STRCAST(val.c_str()), static_cast<int>(val.length()), 1);
     }
     catch(const IceUtil::Exception& ex)
     {
@@ -123,7 +123,7 @@ ZEND_METHOD(Ice_Properties, getPropertyWithDefault)
     try
     {
         string val = _this->getPropertyWithDefault(propName, defaultValue);
-        RETURN_STRINGL(STRCAST(val.c_str()), val.length(), 1);
+        RETURN_STRINGL(STRCAST(val.c_str()), static_cast<int>(val.length()), 1);
     }
     catch(const IceUtil::Exception& ex)
     {
